add test for RouteResource

+++ src/main/java/net/wbz/moba/controlcenter/api/scenario/RouteResource.java
@@ -82,11 +82,14 @@ public class RouteResource {
     @DELETE
     @Path("/{id}")
     public Response delete(@PathParam("id") Long id) {
+        if (routeManager.notExistsById(id)) {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
         boolean deleted = routeManager.deleteRoute(id);
         if (deleted) {
             return Response.noContent().build();
         } else {
-            return Response.status(Response.Status.NOT_FOUND).build();
+            return Response.status(Status.NOT_ACCEPTABLE).build();
         }
     }
 }
